add stam mod key
